Re-register reminder alarms via WorkManager backstop and on clock/timezone change
Task reminders are exact AlarmManager alarms pre-scheduled from JS. Two gaps
could silently drop them while the app is closed:

- Aggressive OEM battery killers (Samsung/Xiaomi/etc.) and some system
  restarts clear pending alarms. The 15-minute WidgetUpdateWorker only
  re-armed the Up Next chain, never the task reminders.
- A timezone or clock change (travel, DST, manual change) left the absolute
  alarm instants out of step with local wall-clock time, with no re-sync
  until the next app open.

Changes:
- Extract the BOOT_COMPLETED re-registration loop into a shared, idempotent
  NotificationBridge.reregisterPersistedReminders(): reads the same persisted
  list syncReminders maintains, skips already-fired reminders, and uses
  FLAG_UPDATE_CURRENT so repeated runs update alarms in place instead of
  duplicating them.
- Call it from WidgetUpdateWorker as a 15-minute backstop.
- Add TimeChangeReceiver (TIMEZONE_CHANGED / TIME_SET) that re-registers the
  alarms and refreshes the Up Next notification, which recomputes against the
  new wall clock and self-heals immediately. Discrete reminders are recomputed
  precisely by the JS reminder engine on the next app foreground.

No new permissions, no foreground service, no battery-optimization exemption.
